+++ scripts/logtalk_tester.ps1
@@ -339,7 +339,7 @@ Function Write-Usage-Help() {
 	Write-Output "  -b bug report server (valid values are github and gitlab; no default; requires -u option)"
 	Write-Output "  -u base URL to generate links to test files (no default)"
 	Write-Output "  -c code coverage report (default is $c)"
-	Write-Output "     (valid values are none and xml)"
+	Write-Output "     (valid values are none, xml, cobertura, and lcov)"
 	Write-Output "  -l directory depth level to look for test sets (default is to recurse into all sub-directories)"
 	Write-Output "     (level 1 means current directory only)"
 	Write-Output "  -e exclude directories matching a regular expression"
@@ -490,6 +490,10 @@ Function Confirm-Parameters() {
 		$script:coverage_goal = $coverage_default_goal
 	} elseif ($c -eq "xml") {
 		$script:coverage_goal = $coverage_xml_goal
+	} elseif ($c -eq "cobertura") {
+		$script:coverage_goal = $coverage_cobertura_goal
+	} elseif ($c -eq "lcov") {
+		$script:coverage_goal = $coverage_lcov_goal
 	} else {
 		Write-Error "Error! Unknown coverage report: $c"
 		Write-Usage-Help
@@ -609,6 +613,8 @@ $format_goal = $format_default_goal
 
 $coverage_default_goal = "true"
 $coverage_xml_goal = "logtalk_load(lgtunit(coverage_report))"
+$coverage_cobertura_goal = "logtalk_load(lgtunit(cobertura_report))"
+$coverage_lcov_goal = "logtalk_load(lgtunit(lcov_report))"
 $coverage_goal = $coverage_default_goal
 
 $results = (Join-Path $pwd "logtalk_tester_logs") -replace '\\', '/'